dev-requirements.txt
aioresponses~=0.7.6
croniter~=1.3.8
watchdog~=2.0.3
mock-open~=1.4.0
mypy~=1.10.1
pre-commit~=2.21.0
polyfactory~=2.9.0
ruff~=0.1.9
PyYAML~=6.0.1
types-PyYAML~=6.0.12.20240311

# Need specific versions
pytest-homeassistant-custom-component==0.13.142
myPyllant==0.8.23

# Versions handled by pytest-homeassistant-custom-component
freezegun
pytest
pytest-asyncio
pytest-cov